+++ R/atom-bu.R
@@ -55,8 +55,11 @@ catr_atom_get_buildings <- function(
     "other" = "other"
   )
 
-  munic <- validate_non_empty_arg(munic)
+  munic <- validate_scalar_arg(munic)
   to <- ensure_null(to)
+  if (!is.null(to)) {
+    to <- validate_scalar_arg(to)
+  }
 
   all <- catr_atom_get_buildings_db_all(
     update_cache = update_cache,