+++ R/ovc-rccoor.R
@@ -31,8 +31,8 @@
 #' }
 catr_ovc_get_rccoor <- function(lat, lon, srs = 4326, verbose = FALSE) {
   # Validate arguments.
-  lat <- validate_non_empty_arg(lat)
-  lon <- validate_non_empty_arg(lon)
+  lat <- validate_coordinate_arg(lat)
+  lon <- validate_coordinate_arg(lon)
 
   srs <- ovc_validate_srs(srs)